Create new function to call weather API
As this learning project is purely front-end work, the API key would be
publically exposed if pushed to github. To help avoid this whilst still
allowing the code to function locally, implemented a new module file,
api-config.js, to store the user API key.
For this commit, the API key will be set to 'Your API Key here'. In the
subsequent commit, this file will be added to the .gitignore so that it
doesn't get updated in future commits.
Any user who plans to use this code, will need to add their own API key
to the config file prior to using the app.