Create function to process API data
New function that accepts the JSON returned by the callAPI function and
returns an object that's a subset of the original JSON dataset.

Additionally:
- In the previous commit it was noted that api-config.js would be added
to .gitignore so that future commits would not accidentally publish the
user API key. Unfortunately, once a file has been committed .gitignore
will no longer ignore that file for future commits. As such, the file
will need to be manually removed from future commits.