Implement new form for location input
Created a new HTML with a single input. Upon submission, the form passes
the user input to our custom call API function, and the returned json is
then passed to our custom data processing function. For now, the
returned data is output to the console but will be updated in future
commits to pass the data to our display controller.
Added custom form validation to the input. The input is for the user to
type a location to request the weather for. As such, only character
values should be needed. The custom JavaScript validation will ensure
that only character values (A-Z, a-z) are accepted by the input field.
All other values are deleted immediately and don't appear within the
display.

Additional Changes:
- As the app expands, more custom modules are being defined. To help
improve organisation, created a components subdirectory within src to
store all of these JavaScript module files.
- As part of the page design, the form won't have a visible label. But
we want to maintain accessibility for users who utilise a screen reader.
To do this, the label has been included within the HTML file and have
defined a `.visually-hidden` CSS class that hides the label from view
but still keeps its position within the DOM so it's accessible to screen
readers.