Add functionality to display weather data
Created several new modules that aid in populating the display:
- create-page-structure.js: Contains functions that create a blank
structure for the page which can then be populated with data.
- display-controller.js: Contains functions that take the data returned
by our API call and uses it to populate the display.
- svg-icons.js: Contains a single function that returns all of the HTML
code for each SVG image the page needs.

Additionally:
- Implemented basic CSS styling to more easily see changes made to the
display.
- Modified the data processor function such that it:
    - Returns the resolved location name in full
    - Returns the hourly forecast for each display
    - Only returns the first 7 days of forecasts instead of 14
- Included additional spans inside some HTML display elements to
simplify a future addition that allows the user to toggle the display
between Celsius and Fahrenheit and between mph and km/h.