Add better error handling for API errors
Improved the error handling of the API call function. Instead of
throwing a new error message to the console (where an average user is
unlikely to see it), the API call function will now print an error
message directly in the UI. Alongwith the error code, there is a brief
description included of what the error represents and what the user can
do to fix it (if it's user error).